refactor: split ContentView.swift into focused subviews (#532)
Extract 7 focused files from the 1811-line ContentView.swift:
- SidebarView.swift — file tree, sidebar search, SidebarEditState
- FileNodeRow.swift — sidebar file/folder row with context menu, rename, create
- EditorAreaView.swift — editor content, CodeEditorView wrapper, drag & drop
- StatusBarView.swift — status bar with git summary, cursor info, terminal toggle
- TerminalBarView.swift — terminal tab bar, search bar, search observer
- GitAndNotificationObserver.swift — ViewModifiers for git/notification/blame/terminal observers
- ContentView+Helpers.swift — extension with session, recovery, blame, diff, close helpers

ContentView.swift is now a thin composition root (240 lines).
No logic changes — pure code movement.
